Some warning and code cleanup.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1998 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/ChunkFile.h
@@ -172,7 +172,7 @@ public:
 			}
 
 			// get size
-			int sz = FileSize - sizeof(SChunkHeader);
+			int sz = (int)(FileSize - sizeof(SChunkHeader));
 			if (Header.ExpectedSize != sz)
 			{
 				fclose(pFile);
@@ -213,7 +213,7 @@ public:
 			SChunkHeader Header;
 			Header.Compress = 0;
 			Header.Revision = _Revision;
-			Header.ExpectedSize = sz;
+			Header.ExpectedSize = (int)sz;
 
 			fwrite(&Header, sizeof(SChunkHeader), 1, pFile);
 			fwrite(buffer, sz, 1, pFile);
